refactor: split index.ts into modules along its real seams
index.ts was 2411 lines. What resisted splitting was never the length,
it was eleven mutable module globals read and written from every
region, so those went first into bridge-state.ts. An exported `let`
would not have worked: an importer binds the value, not the variable,
so a reassignment in one module is invisible in another. They live on a
single exported object instead, chosen over accessor pairs because
about thirty call sites read a singleton several times in one
expression or mutate a field in place, and accessors would have turned
a mechanical rename into a restructure. The two that are only ever
mutated in place stay plain exports.

The rest followed the call graph rather than a filing scheme:

  provider.ts        the pi entry point and its request hooks
  stream-consumer.ts SDK messages into pi stream events
  askclaude.ts       promptAndWait and the tool
  session-sync.ts    the resume-versus-rebuild decision
  isolated-summary.ts compaction and branch-summary takeover
  tools.ts           tool name and argument mapping
  tool-delivery.ts   tool results and steers back to Claude Code
  debug.ts           debug and diag logging

index.ts is 273 lines now and holds only the extension registration.
Nothing under src/ imports it, and the graph has no cycles.

A few symbols moved somewhere other than where they started because
the no-cycle rule decided it: resultErrorText and
logServedContextWindow are used by both the isolated summary and
consumeQuery, so both sit in stream-consumer.ts; turnStart and the two
prompt extractors form a chain syncSharedSession also needs, so the
three went to convert.ts together; askClaudeToolName is reassigned by
the registration and read by both the provider and AskClaude, so it is
state, not a constant.

tool-delivery.ts exists because provider.ts first landed on exactly
600, the ceiling, by trimming a header comment to fit. That is a number
picked to satisfy the rule rather than a seam. The steer-before-result
FIFO ordering is the delicate part of the bridge and reads better
without the query lifecycle wrapped around it; the split leaves 498 and
120.

Pure move: every relocated symbol is byte-identical to its previous
form apart from the added export and the bridgeState prefix. Unit
suite 201/201 and typecheck clean after each extraction, and the
integration suites match their pre-split baseline exactly, down to the
same three failures caused by a missing OpenRouter key.
