+++ src/askclaude.ts
@@ -1,0 +1,408 @@
+// The AskClaude tool: pi delegating a question or a task to a Claude Code
+// subprocess that runs Claude Code's own builtin tools rather than pi's.
+//
+// `promptAndWait` spawns and drains that subprocess; the rest is the pi tool
+// definition wrapped around it — mode presets for which CC tools to block, the
+// TUI renderers, and the registration itself, which only happens when the user
+// has opted in via askClaude.enabled.
+//
+// Separate from index.ts because it shares almost nothing with the provider path:
+// no MCP bridge, no QueryContext, no streaming back into pi. What the two do share
+// is the resume/rebuild decision in session-sync.js, which a non-isolated AskClaude
+// calls so its child resumes the same session file the provider writes. The name
+// this tool is registered under lives on bridgeState, because the provider has to
+// exclude that same name from the pi tools it serves.
+
+import { StringEnum, type Context } from "@earendil-works/pi-ai";
+import { buildSessionContext, keyHint, type ExtensionAPI } from "@earendil-works/pi-coding-agent";
+import { query, type SDKMessage, type SettingSource } from "@anthropic-ai/claude-agent-sdk";
+import { Type } from "typebox";
+import { Text } from "@earendil-works/pi-tui";
+import { buildActionSummary, type ToolCallState } from "./askclaude-ui.js";
+import { bridgeState, promptCaptures } from "./bridge-state.js";
+import { CC_CHILD_ENV, CLAUDE_MD_EXCLUDES, claudeCodeSettings, type Config } from "./config.js";
+import { debug, makeCliDebugOptions } from "./debug.js";
+import { errorMessage } from "./isolated-summary.js";
+import { MODELS, claudeCodeModelId, resolveModel as _resolveModel } from "./models.js";
+import { collectPromptSkills } from "./prompt-capture.js";
+import { syncSharedSession } from "./session-sync.js";
+import { renderSkillsBlock } from "./skills.js";
+import { REASONING_TO_EFFORT, resultErrorText } from "./stream-consumer.js";
+import { mapCustomToolNamesToSdk, mapToolName } from "./tools.js";
+
+function resolveModel(input: string) {
+	return _resolveModel(MODELS, input);
+}
+
+// AskClaude mode presets — controls which CC tools are blocked per mode.
+// Only block tools that can't work (no pi TUI for user interaction).
+// Other CC tools (Agent, SendMessage, RemoteTrigger, Tasks, etc.) are intentionally not blocked.
+const ASKCLAUDE_ALWAYS_BLOCKED = [
+	"AskUserQuestion", "EnterPlanMode", "ExitPlanMode",
+	"ToolSearch", // probes for blocked tools, wastes tokens
+	"ScheduleWakeup", // no harness to fire wakeup from inside a delegated subagent
+];
+const MODE_DISALLOWED_TOOLS: Record<string, string[]> = {
+	full: [
+		...ASKCLAUDE_ALWAYS_BLOCKED,
+	],
+	read: [
+		...ASKCLAUDE_ALWAYS_BLOCKED,
+		"Write", "Edit", "Bash", "NotebookEdit",
+		"EnterWorktree", "ExitWorktree", "CronCreate", "CronDelete", "TeamCreate", "TeamDelete",
+	],
+	none: [
+		...ASKCLAUDE_ALWAYS_BLOCKED,
+		"Read", "Write", "Edit", "Glob", "Grep", "Bash", "Agent",
+		"NotebookEdit", "EnterWorktree", "ExitWorktree",
+		"CronCreate", "CronDelete", "TeamCreate", "TeamDelete",
+		"WebFetch", "WebSearch",
+	],
+};
+
+const DEFAULT_TOOL_DESCRIPTION_FULL = "Delegate to Claude Code for a second opinion or analysis (code review, architecture questions, debugging theories), or to autonomously handle a task. Defaults to read-only mode — use full mode when the user wants to delegate a task that requires changes. Prefer to handle straightforward tasks yourself.";
+const DEFAULT_TOOL_DESCRIPTION = "Delegate to Claude Code for a second opinion or analysis (code review, architecture questions, debugging theories). Read-only — Claude Code can explore the codebase but not make changes. Prefer to handle straightforward tasks yourself.";
+
+const PREVIEW_MAX_CHARS = 1000;
+const PREVIEW_MAX_LINES = 6;
+
+async function promptAndWait(
+	prompt: string,
+	mode: "full" | "read" | "none",
+	toolCalls: Map<string, ToolCallState>,
+	signal?: AbortSignal,
+	options?: {
+		systemPrompt?: string;
+		appendSkills?: boolean;
+		onStreamUpdate?: (responseText: string) => void;
+		model?: string;
+		thinking?: string;
+		isolated?: boolean;
+		context?: Context["messages"];
+		customToolNameToSdk?: Map<string, string>;
+	},
+): Promise<{ responseText: string; stopReason: string }> {
+	const cwd = process.cwd();
+	const requestedModel = options?.model ?? "opus";
+	const model = resolveModel(requestedModel);
+	const modelId = model?.id ?? requestedModel;
+	const cliModel = model ? claudeCodeModelId(model, bridgeState.longContextSettings) : modelId;
+
+	// Session resume for shared mode — reuse provider's session if it exists,
+	// otherwise create one from pi's context.
+	// Note: doesn't update sharedSession.cursor after completion, so the next
+	// provider call will see missed messages and trigger a Case 4 rebuild.
+	// That costs exactly one rebuild, not one per turn: the rebuild sets
+	// cursor = priorMessages.length, and the same turn's tool-result or
+	// query-completion path carries it to the live count, so the turn after
+	// takes REUSE again.
+	// Advancing the cursor here instead would be wrong. CC has appended the
+	// AskClaude exchange as raw user/assistant records while pi represents it
+	// as a toolCall + toolResult, so an advance would make the next turn REUSE
+	// a session file that structurally disagrees with pi's history — the drift
+	// syncSharedSession exists to prevent.
+	let resumeSessionId: string | null = null;
+	if (!options?.isolated && options?.context?.length) {
+		if (bridgeState.sharedSession) {
+			// Provider already has a session — just resume from it
+			// Any missed messages from other providers were already handled by the provider's Case 4
+			resumeSessionId = bridgeState.sharedSession.sessionId;
+		} else {
+			// No provider session yet — create one from pi's context
+			const contextWithPrompt = [...options.context, { role: "user" as const, content: prompt, timestamp: Date.now() }];
+			// isReentrant: true unconditionally — a non-isolated AskClaude is by
+			// construction nested inside a pi turn, so it must never take over the
+			// shared session even when its context happens to be the longer one.
+			const sync = syncSharedSession(contextWithPrompt as Context["messages"], cwd, true, options.customToolNameToSdk, cliModel);
+			resumeSessionId = sync.sessionId;
+		}
+	}
+
+	// Mode → disallowed tools
+	const disallowedTools = MODE_DISALLOWED_TOOLS[mode] ?? [];
+
+	// AskClaude uses Claude Code's native Read tool rather than Pi's MCP bridge.
+	// Same resolver as the provider path: a prompt neither recorded nor derivable
+	// throws here too, rather than silently sending Claude Code no skills.
+	//
+	// Resolved only when the answer would be used. The throw is justified by what a
+	// miss would cost, so where it costs nothing — skills switched off, or no reader
+	// to open a skill file with — an unrelated miss must not fail the call.
+	const skillReadTool = disallowedTools.includes("Read") ? "none" : "native";
+	const skillCapture = options?.appendSkills !== false && skillReadTool !== "none"
+		? promptCaptures.resolveOrDerive(options?.systemPrompt)
+		: undefined;
+	const skillsBlock = skillCapture
+		? renderSkillsBlock(collectPromptSkills(skillCapture), skillReadTool)
+		: undefined;
+
+	// Effort
+	const effort = options?.thinking && options.thinking !== "off"
+		? REASONING_TO_EFFORT[options.thinking] : undefined;
+
+	const claudeExecutable = bridgeState.providerSettings.pathToClaudeCodeExecutable;
+
+	const extraArgs: Record<string, string | null> = {
+		"strict-mcp-config": null,
+		model: cliModel,
+	};
+	if (effort) extraArgs["thinking-display"] = "summarized";
+
+	debug("askClaude:",
+		`mode=${mode} model=${modelId} cliModel=${cliModel} effort=${effort ?? "default"}`,
+		`isolated=${options?.isolated ?? false} resume=${resumeSessionId?.slice(0, 8) ?? "none"}`,
+		`skills=${Boolean(skillsBlock)} promptLen=${prompt.length}`);
+
+	// skills: [] suppresses Claude Code's own skill listing, a system-reminder naming every
+	// skill under the ~/.claude estate. The provider path gets this for free — `tools: []`
+	// removes the Skill tool and the listing with it — but AskClaude runs on CC's native
+	// tools, so it has to be asked for. Pi-side skills still arrive via skillsBlock below,
+	// which is meant to be the only channel.
+	const sdkQuery = query({
+		prompt,
+		options: {
+			cwd,
+			env: { ...process.env, ...CC_CHILD_ENV },
+			permissionMode: "bypassPermissions",
+			// Without this the SDK emits no stream_event messages at all, so the
+			// stream_event case below never runs and onStreamUpdate never fires
+			// during generation — the answer arrives in one lump via the result
+			// fallback, with no live tool-call progress.
+			includePartialMessages: true,
+			settings: { ...claudeCodeSettings(bridgeState.providerSettings), claudeMdExcludes: CLAUDE_MD_EXCLUDES },
+			skills: [],
+			...(disallowedTools.length ? { disallowedTools } : {}),
+			...(effort ? { effort } : {}),
+			// Preset unconditionally: omitting it leaves the child on the SDK's bare default,
+			// without the tool and permission guidance the bridge relies on everywhere else.
+			// Whether pi has skills to append is unrelated to whether the child needs that.
+			systemPrompt: { type: "preset", preset: "claude_code", append: skillsBlock },
+			// "project" only: dropping "user" finishes what claudeMdExcludes and skills: []
+			// started, since ~/.claude/settings.json is also where user-level agents, hooks,
+			// outputStyle and plugins come from — the rest of the estate an AskClaude child
+			// has no business inheriting. Cost of the narrowing: a `env`/`apiKeyHelper` kept
+			// in user settings (Bedrock/Vertex) no longer reaches this child, so auth has to
+			// come from the environment the bridge already forwards.
+			settingSources: ["project"] as SettingSource[],
+			extraArgs,
+			...(resumeSessionId ? { resume: resumeSessionId } : {}),
+			...(options?.isolated ? { persistSession: false } : {}),
+			...(claudeExecutable ? { pathToClaudeCodeExecutable: claudeExecutable } : {}),
+			...makeCliDebugOptions("askclaude"),
+		},
+	});
+
+	// Abort handling
+	let wasAborted = false;
+	const onAbort = () => {
+		wasAborted = true;
+		sdkQuery.interrupt().catch(() => { try { sdkQuery.close(); } catch {} });
+	};
+	if (signal?.aborted) { onAbort(); throw new Error("Aborted"); }
+	signal?.addEventListener("abort", onAbort, { once: true });
+
+	let responseText = "";
+	let sdkMessageCount = 0;
+	let textDeltaCount = 0;
+	let resultSubtype: string | undefined;
+
+	try {
+		for await (const message of sdkQuery) {
+			if (wasAborted) break;
+			sdkMessageCount++;
+
+			switch (message.type) {
+				case "stream_event": {
+					const event = (message as SDKMessage & { event: any }).event;
+					// Text deltas → accumulate and stream
+					if (event?.type === "content_block_delta" && event.delta?.type === "text_delta") {
+						responseText += event.delta.text;
+						textDeltaCount++;
+						options?.onStreamUpdate?.(responseText);
+					}
+					// Tool call start → track for action summary progress
+					if (event?.type === "content_block_start" && event.content_block?.type === "tool_use") {
+						debug(`askClaude: tool_use start: ${event.content_block.name}`);
+						toolCalls.set(event.content_block.id, {
+							name: mapToolName(event.content_block.name),
+							status: "running",
+						});
+					}
+					break;
+				}
+				case "assistant": {
+					// Update tool calls with full input for action summary
+					for (const block of (message as any).message?.content ?? []) {
+						if (block.type === "tool_use") {
+							toolCalls.set(block.id, {
+								name: mapToolName(block.name),
+								status: "complete",
+								rawInput: block.input,
+							});
+						}
+					}
+					break;
+				}
+				case "result": {
+					resultSubtype = message.subtype;
+					const r = message as any;
+					if (r.usage) {
+						debug(`askClaude: result usage: in=${r.usage.input_tokens} out=${r.usage.output_tokens} cacheRead=${r.usage.cache_read_input_tokens ?? 0} cacheWrite=${r.usage.cache_creation_input_tokens ?? 0} turns=${r.num_turns ?? "?"}`);
+					}
+					// Claude Code reports an API failure with `is_error` on a result whose
+					// subtype is still "success", so without this the error text was returned
+					// as Claude's answer and pi's model read a 429 as content. Throwing hands
+					// it to the tool's own catch, which renders it as an error result.
+					const failure = wasAborted ? undefined : resultErrorText(message);
+					if (failure) throw new Error(failure);
+					if (!responseText && message.subtype === "success" && message.result) {
+						responseText = message.result;
+					}
+					break;
+				}
+			}
+		}
+
+		const stopReason = wasAborted ? "cancelled" : "stop";
+		debug(`askClaude: done`,
+			`stopReason=${stopReason} resultSubtype=${resultSubtype ?? "none"}`,
+			`sdkMessages=${sdkMessageCount} textDeltas=${textDeltaCount} responseLen=${responseText.length}`,
+			`toolCalls=${toolCalls.size}`);
+		return { responseText, stopReason };
+	} finally {
+		signal?.removeEventListener("abort", onAbort);
+		sdkQuery.close();
+	}
+}
+
+/** Register the AskClaude tool, when the user has opted into it. Skipped entirely
+ *  otherwise: the tool is opt-in, and an unregistered tool costs pi's model nothing. */
+export function registerAskClaudeTool(pi: ExtensionAPI, config: Config): void {
+	const askConf = config.askClaude;
+	const allowFull = askConf?.allowFullMode !== false;
+	const defaultMode = askConf?.defaultMode ?? "read";
+	const defaultIsolated = askConf?.defaultIsolated ?? false;
+	bridgeState.askClaudeToolName = askConf?.name ?? "AskClaude";
+
+	const modeValues = allowFull ? ["read", "full", "none"] as const : ["read", "none"] as const;
+	let modeDesc = `"read" (default): questions about the codebase — review, analysis, explain. "none": general knowledge only (no file access).`;
+	if (allowFull) modeDesc += ` "full": allows writing and bash execution (careful: runs without feedback to pi).`;
+
+	if (askConf?.enabled) {
+		const askClaudeParams = Type.Object({
+			prompt: Type.String({ description: "The question or task for Claude Code. By default Claude sees the full conversation history. Don't research up front, let Claude explore." }),
+			mode: Type.Optional(StringEnum(modeValues, { description: modeDesc })),
+			model: Type.Optional(Type.String({ description: 'Claude model (e.g. "opus", "sonnet", "haiku", or full ID). Defaults to "opus".' })),
+			thinking: Type.Optional(StringEnum(["off", "minimal", "low", "medium", "high", "xhigh"] as const, { description: "Thinking effort level. Omit to use Claude Code's default." })),
+			isolated: Type.Optional(Type.Boolean({ description: "When true, Claude sees only this prompt (clean session). When false (default), Claude sees the full conversation history." })),
+		});
+		pi.registerTool<typeof askClaudeParams>({
+			name: askConf?.name ?? "AskClaude",
+			label: askConf?.label ?? "Ask Claude Code",
+			description: askConf?.description ?? (allowFull ? DEFAULT_TOOL_DESCRIPTION_FULL : DEFAULT_TOOL_DESCRIPTION),
+			parameters: askClaudeParams,
+			renderCall(args, theme) {
+				let text = theme.fg("mdLink", theme.bold("AskClaude "));
+				const mode = args.mode ?? defaultMode;
+				const tags: string[] = [];
+				if (mode !== defaultMode) tags.push(`mode=${mode}`);
+				if (args.model) tags.push(`model=${args.model}`);
+				if (args.thinking) tags.push(`thinking=${args.thinking}`);
+				if (args.isolated) tags.push("isolated");
+				if (tags.length) text += `${theme.fg("accent", `[${tags.join(", ")}]`)} `;
+				const truncated = args.prompt.length > PREVIEW_MAX_CHARS ? args.prompt.substring(0, PREVIEW_MAX_CHARS) : args.prompt;
+				const lines = truncated.split("\n").slice(0, PREVIEW_MAX_LINES);
+				text += theme.fg("muted", `"${lines.join("\n")}"`);
+				if (args.prompt.length > PREVIEW_MAX_CHARS || args.prompt.split("\n").length > PREVIEW_MAX_LINES) text += theme.fg("dim", " …");
+				return new Text(text, 0, 0);
+			},
+			renderResult(result, { expanded, isPartial }, theme) {
+				if (isPartial) {
+					const status = result.content[0]?.type === "text" ? result.content[0].text : "working...";
+					return new Text(theme.fg("mdLink", "◉ Claude Code ") + theme.fg("muted", status), 0, 0);
+				}
+
+				const details = result.details as { prompt?: string; executionTime?: number; actions?: string; error?: boolean } | undefined;
+				const body = result.content[0]?.type === "text" ? result.content[0].text : "";
+
+				let text = details?.error
+					? theme.fg("error", "✗ Claude Code error")
+					: theme.fg("mdLink", "✓ Claude Code");
+
+				if (details?.executionTime) text += ` ${theme.fg("dim", `${(details.executionTime / 1000).toFixed(1)}s`)}`;
+				if (details?.actions) text += ` ${theme.fg("muted", details.actions)}`;
+
+				if (expanded) {
+					if (details?.prompt) text += `\n${theme.fg("dim", `Prompt: ${details.prompt}`)}`;
+					if (details?.prompt && body) text += `\n${theme.fg("dim", "─".repeat(40))}`;
+					if (body) text += `\n${theme.fg("toolOutput", body)}`;
+				} else {
+					const truncated = body.length > PREVIEW_MAX_CHARS ? body.substring(0, PREVIEW_MAX_CHARS) : body;
+					const lines = truncated.split("\n").slice(0, PREVIEW_MAX_LINES);
+					if (lines.length) text += `\n${theme.fg("toolOutput", lines.join("\n"))}`;
+					if (body.length > PREVIEW_MAX_CHARS || body.split("\n").length > PREVIEW_MAX_LINES) text += `\n${theme.fg("dim", `… (${keyHint("app.tools.expand", "to expand")})`)}`;
+
+				}
+
+				return new Text(text, 0, 0);
+			},
+			async execute(_id, params, signal, onUpdate, ctx) {
+				// Guard: circular delegation
+				if (ctx.model?.baseUrl === "claude-bridge") {
+					debug("askClaude: blocked circular delegation (active provider is claude-bridge)");
+					return {
+						content: [{ type: "text" as const, text: "Error: AskClaude cannot be used when the active provider is claude-bridge — you're already running through Claude Code." }],
+						details: { error: true },
+					};
+				}
+
+				const mode = (params.mode ?? defaultMode) as "full" | "read" | "none";
+				const isolated = params.isolated ?? defaultIsolated;
+				const toolCalls = new Map<string, ToolCallState>();
+				const start = Date.now();
+
+				const progressInterval = setInterval(() => {
+					const elapsed = ((Date.now() - start) / 1000).toFixed(0);
+					const summary = buildActionSummary(toolCalls);
+					const status = summary ? `${elapsed}s — ${summary}` : `${elapsed}s — working...`;
+					onUpdate?.({
+						content: [{ type: "text", text: status }],
+						details: { prompt: params.prompt, executionTime: Date.now() - start },
+					});
+				}, 1000);
+
+				try {
+					const result = await promptAndWait(params.prompt, mode, toolCalls, signal, {
+						systemPrompt: ctx.getSystemPrompt(),
+						appendSkills: askConf?.appendSkills,
+						model: params.model,
+						thinking: params.thinking,
+						isolated,
+						context: isolated ? undefined : buildSessionContext(ctx.sessionManager.getBranch()).messages as Context["messages"],
+						customToolNameToSdk: isolated ? undefined : mapCustomToolNamesToSdk(pi.getActiveTools(), bridgeState.askClaudeToolName),
+					});
+					clearInterval(progressInterval);
+					onUpdate?.({ content: [{ type: "text", text: "" }], details: {} });
+					const executionTime = Date.now() - start;
+					const actions = buildActionSummary(toolCalls);
+
+					const text = actions
+						? `${result.responseText}\n\n[Claude Code actions: ${actions}]`
+						: result.responseText;
+					return {
+						content: [{ type: "text" as const, text }],
+						details: { prompt: params.prompt, executionTime, actions },
+					};
+				} catch (err) {
+					clearInterval(progressInterval);
+					debug(`askClaude error: mode=${mode}, model=${params.model ?? "default"}, isolated=${isolated}, elapsed=${((Date.now() - start) / 1000).toFixed(1)}s, error=`, err);
+					const msg = errorMessage(err);
+					return {
+						content: [{ type: "text" as const, text: `Error: ${msg}` }],
+						details: { prompt: params.prompt, executionTime: Date.now() - start, error: true },
+					};
+				}
+			},
+		});
+	}
+}
