+++ src/convert.ts
@@ -1,9 +1,11 @@
 // Pure pi→Anthropic message conversion helpers.
 // Extracted so they can be tested without pulling in the full extension runtime.
 
-import type { Message as PiMessage } from "@earendil-works/pi-ai";
+import type { Context, ImageContent, Message as PiMessage, TextContent, UserMessage } from "@earendil-works/pi-ai";
+import type { Base64ImageSource, ContentBlockParam } from "@anthropic-ai/sdk/resources";
 import type { ContentBlock, Message as SessionMessage } from "cc-session-io";
 import { pascalCase } from "change-case";
+import { debug } from "./debug.js";
 import { MCP_TOOL_PREFIX } from "./skills.js";
 
 export const PROVIDER_ID = "claude-bridge";
@@ -36,7 +38,7 @@ export function sanitizeToolId(id: string, cache: Map<string, string>): string {
  *    a Claude Code builtin would tell the model a builtin it cannot call is
  *    available and was already used. That is the prompt condition behind the
  *    phantom-call deadlock fixed in 122914dd, and the read direction refuses the
- *    same names for the same reason (piToolNameFor in index.ts).
+ *    same names for the same reason (piToolNameFor in tools.ts).
  *  - **Without a map — the AskClaude path.** CC runs its own tools there, so
  *    builtin names are real, matching mapToolName in the other direction.
  */
@@ -90,7 +92,7 @@ function toolResultContent(
 	return blocks;
 }
 
-/** What convertPiMessages discarded, for the debug line in index.ts. */
+/** What convertPiMessages discarded, for the debug line in session-sync.ts. */
 export type DroppedContent = {
 	thinking: number;
 	abortedTurns: number;
@@ -222,4 +224,79 @@ export function convertPiMessages(
 	}
 
 	return { anthropicMessages, sanitizedIds, dropped };
+}
+
+/** Index of the first message of the current user turn — the trailing run of
+ *  user messages that has not been written into the Claude Code session yet.
+ *  Equals messages.length when the last message is not a user message.
+ *
+ *  Single source of truth for the history/prompt split: everything before this
+ *  index is replayed as session history, everything from it onward becomes the
+ *  prompt. Deriving both halves from one index is what keeps a message from
+ *  landing in both — an extension appending a display-only user message after
+ *  the real one (see issue #34) makes the turn longer than one message. */
+export function turnStart(messages: Context["messages"]): number {
+	let i = messages.length;
+	while (i > 0 && messages[i - 1].role === "user") i--;
+	return i;
+}
+
+/** Extract the current user turn as a prompt string. Returns null if the last message is not a user message. */
+export function extractUserPrompt(messages: Context["messages"]): string | null {
+	const turn = messages.slice(turnStart(messages)) as UserMessage[];
+	if (turn.length === 0) return null;
+	// Drop empties before joining so an all-empty turn still yields "" and trips
+	// the caller's empty-prompt guard rather than sending bare newlines.
+	return turn
+		.map((m) => (typeof m.content === "string" ? m.content : messageContentToText(m.content)))
+		.filter((text) => text)
+		.join("\n");
+}
+
+/** Extract the current user turn as ContentBlockParam[] (preserving images).
+ *  Returns null if no images — caller should fall back to string prompt. */
+export function extractUserPromptBlocks(messages: Context["messages"]): ContentBlockParam[] | null {
+	const turn = messages.slice(turnStart(messages)) as UserMessage[];
+	if (turn.length === 0) return null;
+
+	let hasImage = false;
+	const blocks: ContentBlockParam[] = [];
+	for (const message of turn) {
+		const content: (TextContent | ImageContent)[] = typeof message.content === "string"
+			? [{ type: "text", text: message.content }]
+			: message.content;
+		// Off-type content violates UserMessage's contract, so fail rather than
+		// degrade — but name the shape, since the cause is almost always another
+		// extension appending a malformed message, not this file.
+		if (!Array.isArray(content)) {
+			throw new Error(
+				`extractUserPromptBlocks: user message content must be a string or block array, got ${typeof content} — likely a malformed message from another extension`,
+			);
+		}
+		for (const block of content) {
+			if (block.type === "text" && block.text) {
+				blocks.push({ type: "text", text: block.text });
+			} else if (block.type === "image") {
+				// Guard before logging: data-less image blocks do occur, and reading
+				// .length off the missing field in the debug template would throw
+				// before this check ever runs (template args evaluate unconditionally).
+				if (!block.data || !block.mimeType) {
+					debug(`image block missing data or mimeType, skipping: keys=${Object.keys(block).join(",")}`);
+					continue;
+				}
+				debug(`image block: mimeType=${block.mimeType}, data length=${block.data.length}`);
+				hasImage = true;
+				blocks.push({
+					type: "image",
+					source: {
+						type: "base64",
+						media_type: block.mimeType as Base64ImageSource["media_type"],
+						data: block.data,
+					},
+				});
+			}
+		}
+	}
+	debug(`extractUserPromptBlocks: ${turn.length} msgs in turn, ${blocks.length} blocks, types=${blocks.map((b) => b.type).join(",")}`);
+	return hasImage ? blocks : null;
 }
