+++ src/isolated-summary.ts
@@ -1,0 +1,178 @@
+// The compaction and branch-summary takeover path.
+//
+// Both times pi asks a model to summarize its own history — `/compact` and a
+// rewind or fork-at-point with "summarize" — it hands the job to a `streamFn`.
+// On a bridge model that reaches this extension carrying pi's internal
+// summarization prompt, which no `before_agent_start` ever recorded. Rather than
+// route it through the provider, it runs as its own throwaway Claude Code
+// subprocess: no session, no MCP tools, one turn, nothing to drift out of sync
+// with the live session the provider is caching.
+//
+// Separate from index.ts because it shares nothing with the provider path but the
+// model, and because everything here is reachable from a unit test without
+// registering the extension.
+
+import type { AssistantMessage, AssistantMessageEventStream, Context, Model, SimpleStreamOptions } from "@earendil-works/pi-ai";
+import type { CompactionEntry } from "@earendil-works/pi-coding-agent";
+import { query, type SettingSource } from "@anthropic-ai/claude-agent-sdk";
+import { bridgeState } from "./bridge-state.js";
+import { CC_CHILD_ENV, loadConfig } from "./config.js";
+import { extractUserPrompt } from "./convert.js";
+import { debug, makeCliDebugOptions } from "./debug.js";
+import { claudeCodeModelId } from "./models.js";
+import { logServedContextWindow, newAssistantMessageEventStream, resultErrorText } from "./stream-consumer.js";
+
+export function errorMessage(err: unknown): string {
+	if (err instanceof Error) return err.message;
+	if (err && typeof err === "object") {
+		const obj = err as Record<string, unknown>;
+		if (typeof obj.message === "string") return obj.message;
+		if (typeof obj.error === "string") return obj.error;
+		try { return JSON.stringify(err); } catch {}
+	}
+	return String(err);
+}
+
+export function newAssistantOutput(model: Model<any>, text: string, stopReason: AssistantMessage["stopReason"], errorMessage?: string): AssistantMessage {
+	return {
+		role: "assistant",
+		content: text ? [{ type: "text", text }] : [],
+		api: model.api,
+		provider: model.provider,
+		model: model.id,
+		usage: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, totalTokens: 0,
+			cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 } },
+		stopReason,
+		...(errorMessage ? { errorMessage } : {}),
+		timestamp: Date.now(),
+	};
+}
+
+export function extractIsolatedSummaryPrompt(messages: Context["messages"]): string {
+	if (messages.length !== 1 || messages[0].role !== "user") {
+		throw new Error(
+			`isolatedStreamFn: expected exactly 1 user message, got ${messages.length} ` +
+			`(${messages.map((m) => m.role).join(",")})`,
+		);
+	}
+	const promptText = extractUserPrompt(messages);
+	if (!promptText) throw new Error("isolatedStreamFn: summarization prompt is empty");
+	return promptText;
+}
+
+export function isolatedStreamFn(model: Model<any>, context: Context, options?: SimpleStreamOptions): AssistantMessageEventStream {
+	const stream = newAssistantMessageEventStream();
+	void runIsolatedSummary(model, context, options, stream);
+	return stream;
+}
+
+export async function runIsolatedSummary(
+	model: Model<any>,
+	context: Context,
+	options: SimpleStreamOptions | undefined,
+	stream: AssistantMessageEventStream,
+): Promise<void> {
+	let sdkQuery: ReturnType<typeof query> | undefined;
+	let wasAborted = false;
+	const onAbort = () => {
+		wasAborted = true;
+		void sdkQuery?.interrupt().catch(() => {});
+		try { sdkQuery?.close(); } catch {}
+	};
+
+	try {
+		const promptText = extractIsolatedSummaryPrompt(context.messages);
+		const cwd = (options as { cwd?: string } | undefined)?.cwd ?? process.cwd();
+		const compactProviderSettings = loadConfig(cwd).provider;
+		const claudeExecutable = compactProviderSettings?.pathToClaudeCodeExecutable;
+		const cliModel = claudeCodeModelId(model, bridgeState.longContextSettings);
+		debug(`compact summary: spawn model=${cliModel} registeredModel=${model.id} promptLen=${promptText.length}`);
+
+		sdkQuery = query({
+			prompt: promptText,
+			options: {
+				cwd,
+				env: { ...process.env, ...CC_CHILD_ENV },
+				settings: { autoMemoryEnabled: false },
+				tools: [],
+				strictMcpConfig: true,
+				settingSources: [] as SettingSource[],
+				skills: [],
+				persistSession: false,
+				systemPrompt: context.systemPrompt,
+				model: cliModel,
+				maxTurns: 1,
+				...(claudeExecutable ? { pathToClaudeCodeExecutable: claudeExecutable } : {}),
+				...makeCliDebugOptions("compact-summary"),
+			},
+		});
+
+		if (options?.signal) {
+			if (options.signal.aborted) onAbort();
+			else options.signal.addEventListener("abort", onAbort, { once: true });
+		}
+
+		let assistantText = "";
+		let finalText = "";
+		let errorText: string | undefined;
+		let firstEventLogged = false;
+
+		for await (const message of sdkQuery) {
+			if (!firstEventLogged) {
+				debug(`compact summary: first event type=${message.type}`);
+				firstEventLogged = true;
+			}
+			if (wasAborted) break;
+
+			if (message.type === "assistant") {
+				for (const block of (message as any).message?.content ?? []) {
+					if (block.type === "text" && typeof block.text === "string") assistantText += block.text;
+				}
+			} else if (message.type === "result") {
+				logServedContextWindow("compact summary", message, model);
+				errorText = resultErrorText(message);
+				if (!errorText && message.subtype === "success") finalText = message.result || assistantText;
+			}
+		}
+
+		if (wasAborted) {
+			const output = newAssistantOutput(model, "", "aborted", "Operation aborted");
+			debug("compact summary: aborted");
+			stream.push({ type: "error", reason: "aborted", error: output });
+			stream.end();
+			return;
+		}
+
+		const text = finalText || assistantText;
+		if (errorText || !text.trim()) {
+			const msg = errorText ?? "Claude Code summary returned empty text";
+			debug(`compact summary: error ${msg}`);
+			stream.push({ type: "error", reason: "error", error: newAssistantOutput(model, "", "error", msg) });
+			stream.end();
+			return;
+		}
+
+		debug(`compact summary: done textLen=${text.length}`);
+		stream.push({ type: "done", reason: "stop", message: newAssistantOutput(model, text, "stop") });
+		stream.end();
+	} catch (err) {
+		const msg = errorMessage(err);
+		debug("runIsolatedSummary threw; pushing terminal error", err);
+		stream.push({ type: "error", reason: "error", error: newAssistantOutput(model, "", "error", msg) });
+		stream.end();
+	} finally {
+		options?.signal?.removeEventListener("abort", onAbort);
+		try { sdkQuery?.close(); } catch {}
+	}
+}
+
+export function reinjectPriorCompactionFileOps(branchEntries: Array<{ type: string; details?: unknown }>, preparation: { fileOps: { read: Set<string>; edited: Set<string> } }): void {
+	const prior = [...branchEntries]
+		.reverse()
+		.find((entry): entry is CompactionEntry => entry.type === "compaction");
+	const details = prior?.details as { readFiles?: unknown; modifiedFiles?: unknown } | undefined;
+	if (!Array.isArray(details?.readFiles) || !Array.isArray(details?.modifiedFiles)) return;
+	for (const file of details.readFiles) preparation.fileOps.read.add(String(file));
+	for (const file of details.modifiedFiles) preparation.fileOps.edited.add(String(file));
+	debug(`compact takeover: re-injected prior file ops read=${details.readFiles.length} modified=${details.modifiedFiles.length}`);
+}
