+++ src/session-sync.ts
@@ -1,0 +1,286 @@
+// Session file persistence and the resume/rebuild decision.
+//
+// Everything that writes Claude Code's session JSONL from pi's message history,
+// and `syncSharedSession`, the one function that decides whether a turn can
+// --resume the file that is already there or has to rewrite it from scratch.
+//
+// Separate from index.ts because both callers of that decision — the provider
+// path and a non-isolated AskClaude — need it while neither owns it, and because
+// the rebuild rules are the part of the bridge that unit tests drive hardest
+// (tests/unit-sync-shared-session.mjs). SessionState and the `sharedSession` slot
+// it fills live in bridge-state.js, so nothing here owns the state it advances.
+
+import type { Context } from "@earendil-works/pi-ai";
+import { createSession, deleteSession, openSession, repairToolPairing } from "cc-session-io";
+import { realpathSync, statSync } from "fs";
+import { collectCarriedAttachments, placeCarriedAttachments, type CarriedAttachment } from "./attachments.js";
+import { bridgeState } from "./bridge-state.js";
+import { convertPiMessages, turnStart } from "./convert.js";
+import { DEBUG, DEBUG_LOG_PATH, debug, diagDump } from "./debug.js";
+import { verifyWrittenSession as _verifyWrittenSession } from "./session-verify.js";
+
+/**
+ * Claude Code's `@file` expansions from the session about to be replaced.
+ *
+ * Must be called before `deleteSession`, which wipes the file they live in —
+ * reading after it yields nothing, with no error to notice.
+ */
+function readCarriedAttachments(sessionId: string, cwd: string): CarriedAttachment[] {
+	try {
+		const previous = openSession({ sessionId, projectPath: cwd, claudeDir: process.env.CLAUDE_CONFIG_DIR });
+		return collectCarriedAttachments(previous.records);
+	} catch (error) {
+		// A post-abort rebuild reads a file the killed CC subprocess may have been
+		// midway through writing, and cc-session-io parses each line with a bare
+		// JSON.parse, so a truncated last line throws. Throwing here would turn a
+		// lost attachment into a failed turn; carrying none is exactly what happened
+		// before this existed, so the failure mode is bounded by the status quo.
+		debug(`WARNING: could not read attachments from session ${sessionId.slice(0, 8)}:`, error);
+		return [];
+	}
+}
+
+// Convert pi messages to Anthropic API format for session import.
+// Lossy: only text, thinking and toolCall blocks survive, and thinking only when
+// Claude Code itself minted the signature. An assistant message whose blocks all
+// filter out keeps its slot with a placeholder, since dropping it can create a
+// tool_result with no preceding tool_use. A turn aborted before anything streamed
+// is dropped instead — it never had content, and inventing one diverges from the
+// prefix Claude Code cached.
+function convertAndImportMessages(
+	session: ReturnType<typeof createSession>,
+	messages: Context["messages"],
+	customToolNameToSdk?: Map<string, string>,
+	carried?: readonly CarriedAttachment[],
+): void {
+	const { anthropicMessages, sanitizedIds, dropped } = convertPiMessages(messages, customToolNameToSdk);
+
+	debug(`convertAndImportMessages: ${messages.length} pi msgs → ${anthropicMessages.length} anthropic msgs`);
+	debug(`convertAndImportMessages: imported roles:`, anthropicMessages.map((m, i) => {
+		const c = m.content;
+		if (typeof c === "string") return `[${i}]${m.role}:text`;
+		if (Array.isArray(c)) return `[${i}]${m.role}:${(c).map((b) => b.type).join("+")}`;
+		return `[${i}]${m.role}:?`;
+	}).join(" "));
+	// The roles line above shows only what survived, so a stripped block is
+	// indistinguishable there from one that never existed. Name the losses.
+	const droppedParts = [
+		dropped.thinking ? `${dropped.thinking} thinking (${[...dropped.providers].sort().join(", ")})` : "",
+		dropped.abortedTurns ? `${dropped.abortedTurns} aborted turn(s)` : "",
+		...[...dropped.other].map(([type, n]) => `${n} ${type}`),
+	].filter(Boolean);
+	if (droppedParts.length > 0) {
+		debug(`convertAndImportMessages: dropped ${droppedParts.join(", ")}`);
+	}
+	if (sanitizedIds.size > 0) {
+		debug(`convertAndImportMessages: sanitized ${sanitizedIds.size} tool IDs:`,
+			[...sanitizedIds.entries()].map(([orig, clean]) => orig === clean ? orig : `${orig}→${clean}`).join(", "));
+	}
+	// Pre-repair for debug logging; importMessages also repairs internally (idempotent).
+	const repaired = repairToolPairing(anthropicMessages);
+	if (repaired.length !== anthropicMessages.length) {
+		debug(`convertAndImportMessages: repairToolPairing ${anthropicMessages.length} → ${repaired.length} msgs`);
+	}
+	// Placement runs against the repaired array because that is the index space
+	// importMessages reads. Attachments are links in CC's uuid chain, so they have
+	// to be written in order with the messages, not appended afterwards.
+	const placed = carried?.length
+		? placeCarriedAttachments(carried, repaired as unknown as { role: string; content: unknown }[])
+		: undefined;
+	if (placed?.skipped.length) {
+		debug(`convertAndImportMessages: dropped ${placed.skipped.length} carried attachment(s): ${placed.skipped.join("; ")}`);
+	}
+	if (placed?.attachments.length) {
+		debug(`convertAndImportMessages: carrying ${placed.attachments.length} attachment(s) across the rebuild`);
+	}
+	if (repaired.length) {
+		session.importMessages(repaired, placed?.attachments.length ? { attachments: placed.attachments } : undefined);
+	}
+}
+
+export interface SyncResult {
+	sessionId: string | null;
+	preserveSharedSession?: boolean;
+}
+
+/**
+ * Ensure the shared session has all messages up to (but not including) the last user message.
+ * Returns session ID to resume from, or null if no resume needed.
+ */
+// Read the session file we just wrote and sanity-check it. Warns instead of
+// throwing — CC may be more tolerant than our checks, so a false positive
+// shouldn't block the user. Pure logic is in session-verify.js; this wrapper
+// fans each warning out to debug log + piUI notify + diagDump.
+function verifyWrittenSession(
+	jsonlPath: string,
+	expectedSessionId: string,
+	expectedRecordCount: number,
+	cwd: string,
+): void {
+	const warnings = _verifyWrittenSession(jsonlPath, expectedSessionId, expectedRecordCount);
+	for (const msg of warnings) {
+		debug(`WARNING session verify: ${msg}`);
+		bridgeState.piUI?.notify(
+			`Session file issue: ${msg}\n` +
+			`cwd=${cwd} realpath=${safeRealpath(cwd)} CLAUDE_CONFIG_DIR=${process.env.CLAUDE_CONFIG_DIR ?? "(unset)"}\n` +
+			`Please copy and paste this message into a new issue at https://github.com/elidickinson/pi-claude-bridge/issues/new` +
+			(DEBUG ? ` and attach ${DEBUG_LOG_PATH}` : ` (rerun with CLAUDE_BRIDGE_DEBUG=1 to capture a debug log)`),
+			"warning",
+		);
+		diagDump("session_verify_fail", { msg, jsonlPath, cwd, realpath: safeRealpath(cwd), claudeConfigDir: process.env.CLAUDE_CONFIG_DIR ?? null });
+	}
+}
+
+function safeRealpath(p: string): string {
+	try { return realpathSync(p); } catch (e) { return `<failed: ${(e as Error).message}>`; }
+}
+
+// Diagnostic snapshot of where a session file was just written. Catches the
+// class of bugs where pi writes to ~/.claude/projects/<X> but CC SDK reads
+// from ~/.claude/projects/<Y> (symlinks, CLAUDE_CONFIG_DIR, hash mismatch).
+function debugSessionPaths(label: string, cwd: string, jsonlPath: string): void {
+	const realCwd = safeRealpath(cwd);
+	let fileSize: number | null = null;
+	let fileExists = false;
+	try {
+		const st = statSync(jsonlPath);
+		fileExists = true;
+		fileSize = st.size;
+	} catch { /* file may not exist yet */ }
+	debug(`${label}: cwd=${cwd}`);
+	if (realCwd !== cwd) debug(`${label}: realpath(cwd)=${realCwd} (DIFFERS — symlink-resolved path is what CC SDK uses)`);
+	debug(`${label}: jsonlPath=${jsonlPath}`);
+	debug(`${label}: fileExists=${fileExists}${fileSize != null ? ` size=${fileSize}` : ""}`);
+	debug(`${label}: env.CLAUDE_CONFIG_DIR=${process.env.CLAUDE_CONFIG_DIR ?? "(unset)"} HOME=${process.env.HOME ?? "(unset)"}`);
+}
+
+// Two semantic paths:
+//   REUSE — pi's history is in sync with the existing sharedSession (or drifted
+//     only by the trailing final-assistant message that pi appends after
+//     streamSimple returns, which CC's own persisted session already has).
+//     Returns the existing sessionId. Keeps CC's prompt cache warm.
+//   REBUILD — no session yet, or pi's history has diverged (non-trailing
+//     missed messages, e.g. another provider took a turn). Wipes the existing
+//     session file (if any) and writes a fresh one containing all prior
+//     messages, reusing the same sessionId across rebuilds so UUIDs stay
+//     stable for the lifetime of pi's session.
+//
+// Why a full rebuild rather than patching:
+//   Injecting deltas into an existing session creates a branch that CC's
+//   --resume doesn't follow (documented attempt prior to this). A complete
+//   overwrite at the same path is simpler and correct.
+//
+// Why reuse the sessionId across rebuilds:
+//   CC re-reads the JSONL on every --resume call — no in-process UUID
+//   caching. Validated in tests/exp-session-clear.mjs, including the case
+//   where CC had appended its own tool_use/tool_result records between
+//   rebuilds. Preserving the UUID means stable log correlation across
+//   provider switches and no orphaned session files.
+//
+// Log strings still say "Case 1/2/3/4" so existing diagnostics (int-cache.sh,
+// int-session-resume.mjs) keep grepping the same anchors.
+export function syncSharedSession(
+	messages: Context["messages"],
+	cwd: string,
+	isReentrant: boolean,
+	customToolNameToSdk?: Map<string, string>,
+	modelId?: string,
+): SyncResult {
+	const priorMessages = messages.slice(0, turnStart(messages)); // everything before the current user turn
+
+	// REUSE path
+	//
+	// Guard on priorMessages.length >= cursor: a shorter incoming context cannot
+	// be a continuation of the cached session. This is the general invariant for
+	// pi-side history rewrites such as /compact and session_tree: without it,
+	// missed = [].slice(cursor) can falsely hit REUSE and resume an unrelated
+	// longer CC session. See issue #25.
+	if (bridgeState.sharedSession && !bridgeState.sharedSession.needsRebuild && priorMessages.length >= bridgeState.sharedSession.cursor) {
+		const missed = priorMessages.slice(bridgeState.sharedSession.cursor);
+		const trailingAssistantOnly =
+			missed.length === 1 && (missed[0] as { role?: string }).role === "assistant";
+		if (missed.length === 0 || trailingAssistantOnly) {
+			if (trailingAssistantOnly) {
+				bridgeState.sharedSession = { ...bridgeState.sharedSession, cursor: priorMessages.length, cwd };
+			}
+			debug(`Case 3: ${trailingAssistantOnly ? "advanced cursor past trailing assistant, " : ""}resuming session ${bridgeState.sharedSession.sessionId.slice(0, 8)}, cursor=${bridgeState.sharedSession.cursor}`);
+			debug(`syncResult: path=reuse sessionId=${bridgeState.sharedSession.sessionId} cursor=${bridgeState.sharedSession.cursor}`);
+			return { sessionId: bridgeState.sharedSession.sessionId };
+		}
+	}
+	// This is what keeps a reentrant subagent from taking over the parent's
+	// session: a subagent starts with priors of its own, shorter than the parent's
+	// cursor, so it lands here, gets a fresh session, and the ephemeral session it
+	// captures is deleted once its query completes (see preserveSharedSession in
+	// provider.js's query-completion handler). Remove this branch and a subagent resumes — then
+	// overwrites — the parent's session. The non-isolated AskClaude path reaches it
+	// the same way, and passes isReentrant: true for that reason.
+	//
+	// It is NOT, despite an earlier comment here, the isolated compact-summary
+	// path: runIsolatedSummary never calls syncSharedSession at all.
+	//
+	// Gated on isReentrant, not on the message count alone. A top-level turn with
+	// a shorter context is not a nested query, it is a history rewrite the bridge
+	// did not hear about — a third-party pi extension that prunes pi's messages
+	// array (issue #30). Inferring reentrancy from the count sent those turns to
+	// Claude Code with resume: null and no history at all, and the damage outlived
+	// the turn: with no tool call the cursor update is skipped on the
+	// preserveSharedSession path and stayed stale, so every later turn was
+	// contextless too; with a tool call the tool-result path lowered the cursor to
+	// the pruned length and the next turn REUSEd the never-deleted stale file,
+	// handing back the full un-pruned history. Falling through to REBUILD instead
+	// rewrites the file from the pruned messages and sets the cursor correctly.
+	//
+	// Only reachable when needsRebuild is false — user-facing history rewrites
+	// (/compact, session_tree, /new, fork) always set needsRebuild or clear
+	// sharedSession before the next syncSharedSession call.
+	if (isReentrant && bridgeState.sharedSession && !bridgeState.sharedSession.needsRebuild && priorMessages.length < bridgeState.sharedSession.cursor) {
+		debug(`Case 1 synthetic: clean start for shorter context, preserving shared session ${bridgeState.sharedSession.sessionId.slice(0, 8)}, cursor=${bridgeState.sharedSession.cursor}`);
+		debug(`syncResult: path=clean-start preserve-shared sessionId=${bridgeState.sharedSession.sessionId} cursor=${bridgeState.sharedSession.cursor}`);
+		return { sessionId: null, preserveSharedSession: true };
+	}
+
+	// REBUILD path
+	if (priorMessages.length === 0) {
+		debug(`Case 1: clean start, ${messages.length} total messages`);
+		debug(`syncResult: path=clean-start`);
+		return { sessionId: null };
+	}
+	const previousSessionId = bridgeState.sharedSession?.sessionId;
+	const previousCursor = bridgeState.sharedSession?.cursor ?? 0;
+	// preserveId: rebuild in place (deleteSession + createSession with the
+	// existing UUID), so prompt-cache UUIDs stay stable for log correlation
+	// and for any tools that key off them. Skipped only when there's a
+	// concurrent writer we shouldn't race — see the forceRotate docs on
+	// SessionState in bridge-state.js.
+	const preserveId = previousSessionId !== undefined && !bridgeState.sharedSession?.forceRotate;
+	// Before deleteSession — it wipes the file these live in.
+	const carried = previousSessionId !== undefined ? readCarriedAttachments(previousSessionId, cwd) : [];
+	if (preserveId) {
+		// Wipe prior jsonl + companion dir (no-op if nothing to wipe).
+		deleteSession(previousSessionId!, cwd, process.env.CLAUDE_CONFIG_DIR);
+	}
+	const session = createSession({
+		projectPath: cwd,
+		claudeDir: process.env.CLAUDE_CONFIG_DIR,
+		...(preserveId ? { sessionId: previousSessionId } : {}),
+		...(modelId ? { model: modelId } : {}),
+	});
+	convertAndImportMessages(session, priorMessages, customToolNameToSdk, carried);
+	session.save();
+	// records, not messages: `messages` filters out the attachment records that
+	// carrying an `@file` expansion across a rebuild writes into the same file.
+	verifyWrittenSession(session.jsonlPath, session.sessionId, session.records.length, cwd);
+	bridgeState.sharedSession = { sessionId: session.sessionId, cursor: priorMessages.length, cwd };
+	if (previousSessionId === undefined) {
+		debug(`Case 2: first turn with ${priorMessages.length} prior messages → session ${session.sessionId.slice(0, 8)}, ${session.records.length} records`);
+	} else if (preserveId) {
+		const missedCount = priorMessages.length - previousCursor;
+		debug(`Case 4: ${missedCount} missed messages, ${priorMessages.length} total → rewrote session ${session.sessionId.slice(0, 8)} (same id), ${session.records.length} records`);
+	} else {
+		debug(`Case 4 post-abort: ${priorMessages.length} total → new session ${session.sessionId.slice(0, 8)} (was ${previousSessionId.slice(0, 8)}, rotated to avoid race with orphan writer), ${session.records.length} records`);
+	}
+	debugSessionPaths(`${session.sessionId.slice(0, 8)}`, cwd, session.jsonlPath);
+	debug(`syncResult: path=rebuild sessionId=${session.sessionId} priors=${priorMessages.length} ${previousSessionId === undefined ? "first" : preserveId ? "preserved" : "rotated-post-abort"}`);
+	return { sessionId: session.sessionId };
+}
